# backend/src/ForgeX.Api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForgeX Authoritative API",
    "version": "1.0.0",
    "description": "Deterministic G-code and analytics authority API. G-code responses contain summary evidence without full toolpaths; analytics reports preserve auditable local-rule evidence."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment origin"
    }
  ],
  "paths": {
...
    "/api/v1/gcode/analyze": {
      "post": {
        "operationId": "analyzeGCode",
        "summary": "Analyze an FDM G-code stream",
        "parameters": [
          {
            "name": "bedSizeMm",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 2000,
              "default": 256
            }
          },
          {
            "name": "coordinateOrigin",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "corner",
                "center"
              ],
              "default": "corner"
            }
          },
          {
            "name": "filamentDensityGPerCm3",
            "in": "query",
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "maximum": 20,
              "default": 1.24
            }
          },
          {
            "name": "materialPriceCnyPerKg",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 5000,
              "default": 0
            }
          },
          {
            "name": "nozzleTemperatureMinC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 500,
              "default": 0
            }
          },
          {
            "name": "nozzleTemperatureMaxC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 500,
              "default": 500
            }
          },
          {
            "name": "bedTemperatureMinC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 200,
              "default": 0
            }
          },
          {
            "name": "materialMaxSpeedMmPerSecond",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 5,
              "maximum": 1000,
              "default": 1000
            }
          },
          {
            "name": "materialMaxFlowMm3PerSecond",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0.2,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "name": "machineProfileId",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$",
              "default": "unspecified-machine"
            }
          },
          {
            "name": "materialProfileId",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$",
              "default": "unspecified-material"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-gcode": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 67108864
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authoritative analysis summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GCodeAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/gcode/analyses": {
      "post": {
        "operationId": "createGCodeAnalysisJob",
        "summary": "Persist a G-code stream and queue an authoritative analysis",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "bedSizeMm",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 2000,
              "default": 256
            }
          },
          {
            "name": "coordinateOrigin",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "corner",
                "center"
              ],
              "default": "corner"
            }
          },
          {
            "name": "filamentDensityGPerCm3",
            "in": "query",
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "maximum": 20,
              "default": 1.24
            }
          },
          {
            "name": "materialPriceCnyPerKg",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 5000,
              "default": 0
            }
          },
          {
            "name": "nozzleTemperatureMinC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 500,
              "default": 0
            }
          },
          {
            "name": "nozzleTemperatureMaxC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 500,
              "default": 500
            }
          },
          {
            "name": "bedTemperatureMinC",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 200,
              "default": 0
            }
          },
          {
            "name": "materialMaxSpeedMmPerSecond",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 5,
              "maximum": 1000,
              "default": 1000
            }
          },
          {
            "name": "materialMaxFlowMm3PerSecond",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0.2,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "name": "machineProfileId",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$",
              "default": "unspecified-machine"
            }
          },
          {
            "name": "materialProfileId",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,79}$",
              "default": "unspecified-material"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-gcode": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 67108864
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job queued",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GCodeJobAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "operationId": "getGCodeAnalysisJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durable job snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GCodeJobSnapshotResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/jobs/{id}/events": {
      "get": {
        "operationId": "streamGCodeAnalysisJobEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE progress, heartbeat, and terminal events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/jobs/{id}/cancel": {
      "post": {
        "operationId": "cancelGCodeAnalysisJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current terminal or cancelled snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GCodeJobSnapshotResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/analytics/reports": {
      "post": {
        "operationId": "analyzeAnalyticsReport",
        "summary": "Recompute a deterministic analytics report for shadow comparison",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "C# deterministic analytics report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsAuthorityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/calibration/train": {
      "post": {
        "operationId": "trainCalibrationModel",
        "summary": "Fit a deterministic Theil-Sen time calibration candidate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationTrainingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "C# calibration training result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationAuthorityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
...
      "Problem": {
        "description": "RFC 9457 problem details with stable ForgeX extensions",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "GCodeAnalysisResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "engine",
          "input",
          "profile",
          "parameters",
          "summary",
          "bounds",
          "layers",
          "visualization",
          "claims",
          "pathTypeCounts",
          "warnings",
          "material",
          "risk"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0"
          },
          "engine": {
            "type": "object",
            "required": [
              "version",
              "source"
            ],
            "properties": {
              "version": {
                "type": "string"
              },
              "source": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          "input": {
            "$ref": "#/components/schemas/GCodeInputSummary"
          },
          "profile": {
            "$ref": "#/components/schemas/GCodeProfileSummary"
          },
          "parameters": {
            "type": "object",
            "required": [
              "bedSizeMm",
              "coordinateOrigin",
              "filamentDensityGPerCm3",
              "materialPriceCnyPerKg",
              "nozzleTemperatureMinC",
              "nozzleTemperatureMaxC",
              "bedTemperatureMinC",
              "materialMaxSpeedMmPerSecond",
              "materialMaxFlowMm3PerSecond"
            ],
            "properties": {
              "bedSizeMm": {
                "type": "number"
              },
              "coordinateOrigin": {
                "type": "string",
                "enum": [
                  "corner",
                  "center"
                ]
              },
              "filamentDensityGPerCm3": {
                "type": "number"
              },
              "materialPriceCnyPerKg": {
                "type": "number"
              },
              "nozzleTemperatureMinC": {
                "type": "number"
              },
              "nozzleTemperatureMaxC": {
                "type": "number"
              },
              "bedTemperatureMinC": {
                "type": "number"
              },
              "materialMaxSpeedMmPerSecond": {
                "type": "number"
              },
              "materialMaxFlowMm3PerSecond": {
                "type": "number"
              }
            },
            "additionalProperties": false
          },
          "summary": {
            "type": "object",
            "required": [
              "totalLayers",
              "heightMm",
              "extrusionLengthMm",
              "travelLengthMm",
              "estimatedTimeSeconds",
              "volumeCm3",
              "filamentLengthM",
              "filamentMassG"
            ],
            "properties": {
              "totalLayers": {
                "type": "integer"
              },
              "heightMm": {
                "type": "number"
              },
              "extrusionLengthMm": {
                "type": "number"
              },
              "travelLengthMm": {
                "type": "number"
              },
              "estimatedTimeSeconds": {
                "type": "number"
              },
              "volumeCm3": {
                "type": "number"
              },
              "filamentLengthM": {
                "type": "number"
              },
              "filamentMassG": {
                "type": "number"
              }
            },
            "additionalProperties": false
          },
          "bounds": {
            "type": "object",
            "required": [
              "minX",
              "maxX",
              "minY",
              "maxY"
            ],
            "properties": {
              "minX": {
                "type": "number"
              },
              "maxX": {
                "type": "number"
              },
              "minY": {
                "type": "number"
              },
              "maxY": {
                "type": "number"
              }
            },
            "additionalProperties": false
          },
          "layers": {
            "type": "array",
            "maxItems": 20000,
            "items": {
              "$ref": "#/components/schemas/GCodeLayerSummary"
            }
          },
          "claims": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "pathTypeCounts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "code",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "additionalProperties": false
            }
          },
          "visualization": {
            "$ref": "#/components/schemas/GCodeToolpathVisualization"
          },
          "material": {
            "$ref": "#/components/schemas/GCodeMaterialEstimate"
          },
          "risk": {
            "$ref": "#/components/schemas/GCodeRiskAssessment"
          }
        },
        "additionalProperties": false
      },
      "GCodeJobAcceptedResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "jobId",
          "status",
          "input",
          "links"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0"
          },
          "jobId": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "degraded",
              "failed",
              "cancelled"
            ]
          },
          "input": {
            "$ref": "#/components/schemas/GCodeInputSummary"
          },
          "links": {
            "$ref": "#/components/schemas/GCodeJobLinks"
          }
        },
        "additionalProperties": false
      },
...
      "AnalyticsReportRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schemaVersion",
          "question",
          "rows",
          "provenance"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0"
          },
          "question": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "rows": {
            "type": "array",
            "minItems": 1,
            "maxItems": 5000,
            "items": {
              "$ref": "#/components/schemas/AnalyticsRow"
            }
          },
          "provenance": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AnalyticsProvenance"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "AnalyticsAuthorityResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schemaVersion",
          "engine",
          "report"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0"
          },
          "engine": {
            "$ref": "#/components/schemas/AnalyticsAuthorityEngine"
          },
          "report": {
            "$ref": "#/components/schemas/AnalyticsReport"
          }
        }
      },
...
      "CalibrationAuthorityResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaVersion", "engine", "training"],
        "properties": {
          "schemaVersion": { "type": "string", "const": "1.0" },
          "engine": {
            "type": "object",
            "additionalProperties": false,
            "required": ["name", "version"],
            "properties": {
              "name": { "type": "string", "const": "forgex-calibration-csharp" },
              "version": { "type": "string" }
            }
          },
          "training": { "type": "object" }
        }
      },
...
    }
  }
}